feat: switch transaction ids and logs igs to uint64 (#910)

+++ internal/api/bulking/handler_stream_json_test.go
@@ -65,7 +65,7 @@ func TestBulkStreamHandlerJSON(t *testing.T) {
 				select {
 				case receive <- BulkElementResult{
 					Data:      ledger.CreatedTransaction{},
-					LogID:     id + 1,
+					LogID:     uint64(id) + 1,
 					ElementID: id,
 				}:
 				case <-time.After(100 * time.Millisecond):
